+++ History.md
@@ -17,6 +17,7 @@
 - When creating a motion, the confirmation page now shows a preview of the generated PDF.
 - In the initiator form, the resolution date for organizations submitting a motion is now optional. An additional optional field to add one's gender was added.
 - When exporting the proposed procedure list into ODS, there is now an option to also include the comments.
+- The admin interface for adding/removing supporters of a motion/amendment now has a function to copy the full list of supporters to the clipboard in a format suitable to paste it into the full-text field later on (to easily transfer the supporter list from one motion to a new one).
 - Bugfix: When a motion type was created from the scratch and motion sections positioned right were added, the layout didn't switch to two-column mode.
 - Bugfix: When an optional image was not uploaded, the LaTeX-based PDF export did not work.
 - Bugfix: It was not possible to delete admin comments in the proposed procedure.